+++ test/test_sig.t
@@ -19,6 +19,10 @@
   >   second : 'b;
   > }
   > [@@deriving jsonschema]
+  > 
+  > type foo = { bar : bar option }
+  > and bar = { foo : foo option }
+  > [@@deriving jsonschema]
   > EOF
   $ ./pp.exe -deriving-keep-w32 both --intf test_sig.mli -o test_sig_out.mli && cat test_sig_out.mli
   type event = {
@@ -44,4 +48,11 @@
   include
     sig
       val pair_jsonschema : Yojson.Basic.t -> Yojson.Basic.t -> Yojson.Basic.t
+    end[@@ocaml.doc "@inline"][@@merlin.hide ]
+  type foo = {
+    bar: bar option }
+  and bar = {
+    foo: foo option }[@@deriving jsonschema]
+  include
+    sig val foo_jsonschema : Yojson.Basic.t val bar_jsonschema : Yojson.Basic.t
     end[@@ocaml.doc "@inline"][@@merlin.hide ]