+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.0.5
+
+- add support for recursive types (self-referential)
+- add support for mutually recursive types (defined with `and`)
+- add support to polymorphic types
+
 # 0.0.4
 
 - add [@@json.allow\_extra\_fields] attribute